+++ module-liquec/src/main/java/edu/uoc/geopocket/liquec/dtos/SptResultDTO.java
@@ -5,23 +5,25 @@ import edu.uoc.geopocket.common.dtos.AuditDTO;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class SptResultDTO {
     private Long id;
     private Result result;
     private String message;
-    private Double totalStress;
-    private Double interstitialStress;
-    private Double effectiveStress;
-    private Double effectiveStressFactor;
-    private Double earthquakeMagnitudeCorrection;
-    private Double coefficientContributionCorrection;
-    private Double sptCorrected;
-    private Double cycleStressRatio;
-    private Double depthFactor;
-    private Double cycleResistanceRatio;
-    private Double cycleResistanceRatioCorrected;
-    private Double safetyFactor;
+    private BigDecimal totalStress;
+    private BigDecimal interstitialStress;
+    private BigDecimal effectiveStress;
+    private BigDecimal effectiveStressFactor;
+    private BigDecimal earthquakeMagnitudeCorrection;
+    private BigDecimal coefficientContributionCorrection;
+    private BigDecimal sptCorrected;
+    private BigDecimal cycleStressRatio;
+    private BigDecimal depthFactor;
+    private BigDecimal cycleResistanceRatio;
+    private BigDecimal cycleResistanceRatioCorrected;
+    private BigDecimal safetyFactor;
     private AuditDTO audit;
 }
